fix: padrão heap JVM definido para mais arquivos no banco de dados e correção na validação do forms
